+++ CHANGES.md
@@ -17,6 +17,11 @@
     `Irmin_pack.Indexing_strategy` and the rest of `Pack_store`
     to `Irmin_pack_unix` (#1833, @Ngoguey42)
 
+### Fixed
+
+- **irmin-pack**
+  - Allow snapshot export to work on indexed root nodes (#1845, @icristescu)
+
 ## 3.2.1 (2022-04-07)
 
 - Support all version of cmdliner (#1803, @samoht)